fix: [ANDROAPP-4845] Move QRDetailBottomDialog to form module
