+++ form/src/main/java/org/dhis2/form/ui/dialog/QRDetailBottomDialog.kt
@@ -1,4 +1,4 @@
-package org.dhis2.utils.customviews
+package org.dhis2.form.ui.dialog
 
 import android.content.Context
 import android.content.Intent
@@ -21,11 +21,11 @@ import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import java.io.File
 import java.io.FileOutputStream
 import java.io.IOException
-import org.dhis2.Bindings.clipWithRoundedCorners
-import org.dhis2.R
+import org.dhis2.commons.bindings.clipWithRoundedCorners
 import org.dhis2.commons.resources.ColorUtils
-import org.dhis2.databinding.QrDetailDialogBinding
+import org.dhis2.form.R
 import org.dhis2.form.data.FormFileProvider
+import org.dhis2.form.databinding.QrDetailDialogBinding
 import org.dhis2.form.model.UiRenderType
 import org.hisp.dhis.android.core.arch.helpers.FileResourceDirectoryHelper
 import timber.log.Timber
