+++ form/src/main/java/org/dhis2/form/ui/dialog/QRImageViewModel.kt
@@ -1,4 +1,4 @@
-package org.dhis2.utils.customviews
+package org.dhis2.form.ui.dialog
 
 import android.graphics.Bitmap
 import androidx.lifecycle.LiveData
