+++ form/src/main/res/drawable/ic_qr_scan.xml
