+++ form/src/main/res/layout/qr_detail_dialog.xml
