+++ form/src/main/res/values-ar/strings.xml
@@ -16,4 +16,5 @@
     <string name="unique_warning">يجب أن تكون القيمة فريدة، يوجد تعارض تكرار لنفس القيمة. لم يتم الحفظ. يرجى التحقق.</string>
     <string name="select_time">إختر الوقت</string>
     <string name="choose_date">اختر التاريخ</string>
+    <string name="share">مشاركة</string>
 </resources>
