+++ form/src/main/res/values-b+uz+Cyrl/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="share">Баҳам кўриш</string>
+</resources>
