+++ form/src/main/res/values-km/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="share">ចែករំលែក</string>
+</resources>
