+++ form/src/main/res/values-lo/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="share">ແລກປ່ຽນ</string>
+</resources>
