+++ form/src/main/res/values-tet/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="share">Partilla</string>
+</resources>
