+++ form/src/main/res/values-uz/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="share">Улашиш</string>
+</resources>
