+++ form/src/main/res/values-zh/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="share">分享</string>
+</resources>
