+++ form/src/main/res/values/strings.xml
@@ -64,5 +64,6 @@
     <string name="field_errors_not_saved">Some fields have errors and they are not saved.</string>
     <string name="fields_mandatory_missing">Some mandatory fields are missing and the form cannot be saved.</string>
     <string name="fields_mandatory_missing_discard">Some mandatory fields are missing and the form cannot be saved. \nIf you exit now the changes will be discarded.</string>
+    <string name="share">Share</string>
 
 </resources>
